Moved advanced options to a new settings tab and linted.
