+++ src/components/ContainerDetailsSubheader.react.js
@@ -1,16 +1,9 @@
-var $ = require('jquery');
 var _ = require('underscore');
 var React = require('react');
-var exec = require('exec');
 var shell = require('shell');
 var metrics = require('../utils/MetricsUtil');
 var ContainerUtil = require('../utils/ContainerUtil');
-var util = require('../utils/Util');
-var machine = require('../utils/DockerMachineUtil');
-var RetinaImage = require('react-retina-image');
 var classNames = require('classnames');
-var resources = require('../utils/ResourcesUtil');
-var dockerUtil = require('../utils/DockerUtil');
 var containerActions = require('../actions/ContainerActions');
 var dockerMachineUtil = require('../utils/DockerMachineUtil');
