+++ src/components/ContainerSettingsGeneral.react.js
@@ -22,14 +22,10 @@ var ContainerSettingsGeneral = React.createClass({
       return [util.randomId(), e[0], e[1]];
     });
 
-    let [tty, openStdin] = ContainerUtil.mode(this.props.container) || [true, true];
-
     return {
       slugName: null,
       nameError: null,
-      env: env,
-      tty: tty,
-      openStdin: openStdin
+      env: env
     };
   },
 
@@ -94,9 +90,7 @@ var ContainerSettingsGeneral = React.createClass({
         list.push(key + '=' + value);
       }
     });
-    let tty = this.state.tty;
-    let openStdin = this.state.openStdin;
-    containerActions.update(this.props.container.Name, {Env: list, Tty: tty, OpenStdin: openStdin});
+    containerActions.update(this.props.container.Name, {Env: list});
   },
 
   handleChangeEnvKey: function (index, event) {
@@ -137,18 +131,6 @@ var ContainerSettingsGeneral = React.createClass({
     });
 
     metrics.track('Removed Environment Variable');
-  },
-
-  handleChangeTty: function () {
-    this.setState({
-      tty: !this.state.tty
-    });
-  },
-
-  handleChangeOpenStdin: function () {
-    this.setState({
-      openStdin: !this.state.openStdin
-    });
   },
 
   handleDeleteContainer: function () {
@@ -229,14 +211,6 @@ var ContainerSettingsGeneral = React.createClass({
           <div className="env-vars">
             {vars}
           </div>
-        </div>
-        <div className="settings-section">
-          <div className="env-vars">
-            <h4>Advanced Options</h4>
-            <p><input type="checkbox" checked={this.state.tty} onChange={this.handleChangeTty}/> Attach standard streams to a tty, including stdin if it is not closed</p>
-            <p><input type="checkbox" checked={this.state.openStdin} onChange={this.handleChangeOpenStdin}/> Keep STDIN open even if not attached</p>
-          </div>
-          <a className="btn btn-action" disabled={this.props.container.State.Updating} onClick={this.handleSaveEnvVars}>Save</a>
         </div>
         <div className="settings-section">
           <h3>Delete Container</h3>
