+++ src/stores/ContainerStore.js
@@ -1,9 +1,7 @@
 import _ from 'underscore';
-import deepExtend from 'deep-extend';
 import alt from '../alt';
 import containerServerActions from '../actions/ContainerServerActions';
 import containerActions from '../actions/ContainerActions';
-var LogStore = require('./LogStore');
 
 class ContainerStore {
   constructor () {
