+++ src/utils/RegHubUtil.js
@@ -5,7 +5,6 @@ var util = require('../utils/Util');
 var hubUtil = require('../utils/HubUtil');
 var repositoryServerActions = require('../actions/RepositoryServerActions');
 var tagServerActions = require('../actions/TagServerActions');
-var Promise = require('bluebird');
 
 let REGHUB2_ENDPOINT = process.env.REGHUB2_ENDPOINT || 'https://registry.hub.docker.com/v2';
 let searchReq = null;
@@ -62,7 +61,7 @@ module.exports = {
       let data = JSON.parse(body);
       let repos = data.repos;
       async.map(repos, (repo, cb) => {
-        let name = repo.repo;
+        var name = repo.repo;
         if (util.isOfficialRepo(name)) {
           name = 'library/' + name;
         }
