+++ styles/preferences.less
@@ -16,7 +16,6 @@
 
     .title {
       margin-top: 40px;
-      border-bottom: 1px solid #EEE;
       text-align: left;
       font-size: 18px;
       font-weight: 400;
